refactor: split exercise session screen
